[ASL] Introduce "skipFeatures" hack (a gross hack!)
It is used to specify that some features (EStructuralFeatures) from the
CS model should not be added to the AS model. Additionally, it is used
to specify that for some features no containment visitor code should be
generated (Note: the cs2as and containment visitor code "skipFeatures"
lists are independent!).
Again, this is a gross hack, and should be fixed in a more "sane" way.